Fixed Delete (No Plan For Deleting Works) Client/Server

+++ src/main/java/hassan/personnel/managment/rests/PositionsController.java
@@ -1,5 +1,7 @@
 package hassan.personnel.managment.rests;
 
+import hassan.personnel.managment.exceptionalResponses.ConflictException;
+import hassan.personnel.managment.exceptionalResponses.NotFoundException;
 import hassan.personnel.managment.models.entities.Position;
 import hassan.personnel.managment.models.entities.Wage;
 import hassan.personnel.managment.models.dto.PositionInsertDto;
@@ -8,9 +10,13 @@ import hassan.personnel.managment.services.PositionService;
 import hassan.personnel.managment.services.WageService;
 import hassan.personnel.managment.utility.CalendarHelper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
 /**
@@ -19,6 +25,8 @@ import java.util.stream.Collectors;
 @RestController
 @RequestMapping("/rest/positions")
 public class PositionsController {
+
+    final static org.apache.log4j.Logger LOGGER = org.apache.log4j.Logger.getLogger(WorksController.class);
 
     @Autowired
     private PositionService positionService;
@@ -47,8 +55,27 @@ public class PositionsController {
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-    private Position remove(@PathVariable int id){
-        return positionService.remove(id);
+    @Transactional(rollbackFor = {Exception.class})
+    private PositionVm remove(@PathVariable int id) throws NotFoundException, ConflictException {
+        Position position = null;
+        LOGGER.warn("THIS METHOD IS NOT TESTED");
+        try {
+            //Check if There are person linked to this Position then throw Conflict(409) Error
+            if(!positionService.getPosition(id).getPersonnel().isEmpty())
+                throw new ConflictException("There are personnel that are linked to this entity");
+
+            //Try to remove all wages linked to this Position
+            for (Wage wage : positionService.getPosition(id).getWages()) {
+                wageService.remove(wage.getId());
+            }
+            //After Deleting All Wages, Delete Position Itself
+            position = positionService.remove(id);
+            return position.getViewModel();
+        }catch (DataIntegrityViolationException ex){
+            throw new ConflictException(ex.getMessage());
+        }catch (EmptyResultDataAccessException ex){
+            throw new NotFoundException("Requested Item Does Not Found");
+        }
     }
 
     @RequestMapping(value = "/fetch-wage", method = RequestMethod.GET)
